Updating with Humphery Clerx patches, fixing misspellings, empty string detections,and adding a node at the end of a poly in edit mode.
git-svn-id: svn+ssh://src.dsl.bbn.com/svn/openmap/trunk@2065 f68c23cc-7f87-4fbc-acfa-d9760b90acc1

+++ src/openmap/com/bbn/openmap/layer/mif/OMSubtraction.java
@@ -104,7 +104,7 @@ public class OMSubtraction extends OMGraphicAdapter implements Serializable {
      * @param y Y coordinate fo the point.
      * @return float always zero
      */
-    public float distance(int x, int y) { //return zero
+    public float distance(double x, double y) { //return zero
         return 0.0f;
     }
