+++ src/openmap/com/bbn/openmap/omGraphics/geom/PolylineGeometry.java
@@ -73,7 +73,7 @@ public abstract class PolylineGeometry extends PolygonGeometry implements
      * @param y vertical pixel location.
      * @return the distance of the object to the location given.
      */
-    public float distance(int x, int y) {
+    public float distance(double x, double y) {
         if (shape != null) {
             return super.distance(x, y);
         }
